moved tx response parser to multiversx-sc-snippets
