+++ framework/scenario/src/scenario/model/transaction.rs
@@ -9,7 +9,6 @@ mod tx_interpret_util;
 mod tx_query;
 mod tx_response;
 mod tx_response_status;
-mod tx_response_utils;
 mod tx_transfer;
 mod tx_validator_reward;
 mod typed_response;
@@ -24,7 +23,6 @@ pub use tx_expect::*;
 pub use tx_query::*;
 pub use tx_response::TxResponse;
 pub use tx_response_status::TxResponseStatus;
-pub use tx_response_utils::*;
 pub use tx_transfer::*;
 pub use tx_validator_reward::*;
 pub use typed_response::TypedResponse;
