+++ framework/snippets/Cargo.toml
@@ -32,3 +32,6 @@ path = "../../sdk/scenario-format"
 [dependencies.multiversx-sdk]
 version = "=0.4.1"
 path = "../../sdk/core"
+
+[dev-dependencies]
+serde_json = "1.0"
