+++ framework/snippets/src/lib.rs
@@ -7,6 +7,7 @@ mod interactor_sender;
 mod interactor_tx;
 mod multi;
 pub mod test_wallets;
+pub mod tx_response_from_network;
 
 pub use env_logger;
 pub use hex;
